Move common computed properties and dialogs to components
